+++ src/main/java/com/group6/wishshare/data/repository/UserRepository.java
@@ -1,7 +1,7 @@
 package com.group6.wishshare.data.repository;
 
 import com.group6.wishshare.domain.model.User;
-import com.group6.wishshare.domain.service.LoginException;
+import com.group6.wishshare.domain.exception.LoginException;
 
 public interface UserRepository {
