+++ src/main/java/com/group6/wishshare/data/repository/mock/UserRepositoryStub.java
@@ -2,7 +2,7 @@ package com.group6.wishshare.data.repository.mock;
 
 import com.group6.wishshare.data.repository.UserRepository;
 import com.group6.wishshare.domain.model.User;
-import com.group6.wishshare.domain.service.LoginException;
+import com.group6.wishshare.domain.exception.LoginException;
 
 public class UserRepositoryStub implements UserRepository {
