+++ src/main/java/com/group6/wishshare/domain/service/UserService.java
@@ -1,6 +1,7 @@
 package com.group6.wishshare.domain.service;
 
 import com.group6.wishshare.data.repository.UserRepository;
+import com.group6.wishshare.domain.exception.LoginException;
 import com.group6.wishshare.domain.model.User;
 import com.group6.wishshare.domain.model.type.Gender;
