+++ src/main/java/com/group6/wishshare/domain/service/WishlistService.java
@@ -9,10 +9,10 @@ import java.util.Random;
 
 public class WishlistService {
 
-  WishlistRepository wishlistRepository;
+  private final WishlistRepository WISHLIST_REPOSITORY;
 
-  public WishlistService(WishlistRepository wishlistRepository) {
-    this.wishlistRepository = wishlistRepository;
+  public WishlistService(WishlistRepository WISHLIST_REPOSITORY) {
+    this.WISHLIST_REPOSITORY = WISHLIST_REPOSITORY;
   }
 
   private String getRandomString() {
@@ -29,30 +29,30 @@ public class WishlistService {
   }
 
   public Wishlist addWishList(User user, String name) {
-    return wishlistRepository.addWishlist(name, user, getRandomString());
+    return WISHLIST_REPOSITORY.addWishlist(name, user, getRandomString());
   }
 
   public Wishlist getWishlistById(int id) {
-    return wishlistRepository.findWishlistById(id);
+    return WISHLIST_REPOSITORY.findWishlistById(id);
   }
 
   public Wishlist findWishListByToken(String token) {
-    return wishlistRepository.findWishlistByToken(token);
+    return WISHLIST_REPOSITORY.findWishlistByToken(token);
   }
 
   public List<Wishlist> getWishlists(User user) {
-    return wishlistRepository.getWishlists(user);
+    return WISHLIST_REPOSITORY.getWishlists(user);
   }
 
   public boolean isOwner(int id, User user) {
-    return wishlistRepository.isWishlistOwnedByUser(id, user);
+    return WISHLIST_REPOSITORY.isWishlistOwnedByUser(id, user);
   }
 
   public void updateWishlistName(int id, String newName) {
-    wishlistRepository.updateName(id, newName);
+    WISHLIST_REPOSITORY.updateName(id, newName);
   }
 
   public void deleteWishlist(int id) {
-    wishlistRepository.deleteWishlist(id);
+    WISHLIST_REPOSITORY.deleteWishlist(id);
   }
 }
