+++ src/main/java/com/group6/wishshare/web/controller/LoginController.java
@@ -3,7 +3,7 @@ package com.group6.wishshare.web.controller;
 import com.group6.wishshare.data.repository.mysql.UserRepositoryImpl;
 import com.group6.wishshare.domain.model.User;
 import com.group6.wishshare.domain.model.type.Gender;
-import com.group6.wishshare.domain.service.LoginException;
+import com.group6.wishshare.domain.exception.LoginException;
 import com.group6.wishshare.domain.service.UserService;
 import com.group6.wishshare.web.util.SessionObject;
 import org.springframework.stereotype.Controller;
@@ -29,7 +29,7 @@ public class LoginController {
 
   @GetMapping("/login")
   public String getLogin() {
-    return "login";
+    return "index";
   }
 
   @GetMapping("/signup")
@@ -45,10 +45,10 @@ public class LoginController {
 
       User user = USER_SERVICE.login(email, password);
       new SessionObject(webRequest).setUser(user);
-      return "redirect:dashboard";
+      return "redirect:/dashboard";
     } catch (LoginException e) {
       model.addAttribute("loginFail", "Wrong email or password"); // TODO
-      return "login";
+      return "index";
     }
   }
 
@@ -73,7 +73,7 @@ public class LoginController {
                 email,
                 password1);
         new SessionObject(webRequest).setUser(user);
-        return "redirect:dashboard";
+        return "redirect:/dashboard";
       }
       model.addAttribute("signupFail", "The passwords did not match"); // TODO
       return "signup";
@@ -86,7 +86,7 @@ public class LoginController {
   @GetMapping("/logout")
   public String logout(WebRequest webRequest) {
     new SessionObject(webRequest).removeUser();
-    return "redirect:";
+    return "redirect:/";
   }
 
   private boolean validatePassword(String password1, String password2) { // TODO move
