--- src/main/resources/templates/login.html
@@ -1,75 +1,0 @@
-<!DOCTYPE html>
-<html class="html-form" lang="en" xmlns:th="http://www.thymeleaf.org">
-
-<head>
-    <div th:replace="fragments/head"></div>
-    <title>
-        Login
-    </title>
-</head>
-
-<body class="">
-<div th:replace="fragments/navbar"></div>
-<main class="main-content mt-0">
-    <section>
-        <div class="page-header min-vh-75">
-            <div class="container">
-                <div class="row">
-                    <div class="col-xl-4 col-lg-5 col-md-6 d-flex flex-column mx-auto">
-                        <div class="card card-plain mt-8">
-                            <div class="card-header pb-0 text-left bg-transparent">
-                                <h3 class="font-weight-bolder text-info text-gradient">Welcome back</h3>
-                                <p class="mb-0">Enter your email and password to sign in</p>
-                            </div>
-                            <div class="card-body">
-                                <form action="/login" method="POST" name="login" role="form">
-                                    <label>Email</label>
-                                    <div class="mb-3">
-                                        <input aria-describedby="email-addon" aria-label="Email" class="form-control"
-                                               name="email"
-                                               placeholder="Email"
-                                               required
-                                               type="email">
-                                    </div>
-                                    <label>Password</label>
-                                    <div class="mb-3">
-                                        <input aria-describedby="password-addon" aria-label="Password"
-                                               class="form-control"
-                                               name="password"
-                                               placeholder="Password"
-                                               required
-                                               type="password">
-                                    </div>
-                                    <div class="form-check form-switch">
-                                        <input checked="" class="form-check-input" id="rememberMe" type="checkbox">
-                                        <label class="form-check-label" for="rememberMe">Remember me</label>
-                                    </div>
-                                    <div class="text-center">
-                                        <p class="mb-0" th:text="${loginFail}"></p>
-                                    </div>
-                                    <div class="text-center">
-                                        <input class="btn bg-gradient-info w-100 mt-4 mb-0" type="submit"
-                                               value="Sign in">
-                                    </div>
-                                </form>
-                            </div>
-                            <div class="card-footer text-center pt-0 px-lg-2 px-1">
-                                <p class="mb-4 text-sm mx-auto">
-                                    Don't have an account?
-                                    <a class="text-info text-gradient font-weight-bold" href="/signup">Sign up</a>
-                                </p>
-                            </div>
-                        </div>
-                    </div>
-                    <div class="col-md-6">
-                        <div class="oblique position-absolute top-0 h-100 d-md-block d-none me-n8">
-                            <div class="oblique-image bg-cover position-absolute fixed-top ms-auto h-100 z-index-0 ms-n6"
-                                 th:style="'background-image: url(/img/curved-images/curved6.jpg)'">
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-</main>
